chore: add basic test
